refactor: remove duplication on speakerAdapter.ts
