+++ src/views/Speakers/UseFetchSpeakers.ts
@@ -1,6 +1,7 @@
-import { useQuery, UseQueryResult } from "react-query";
+import {useQuery, UseQueryResult} from "react-query";
 import axios from "axios";
-import { IResponse, ISpeaker } from "./Speaker.types";
+import {ISpeaker} from "./Speaker.types";
+import {speakerAdapter} from "../../services/speakerAdapter";
 
 export const useFetchSpeakers = (id?: string): UseQueryResult<ISpeaker[]> => {
   return useQuery("api-speakers", async () => {
@@ -18,16 +19,4 @@ export const useFetchSpeakers = (id?: string): UseQueryResult<ISpeaker[]> => {
     return speakerAdapter(returnData);
   });
 };
-export const speakerAdapter = (response: IResponse[]): ISpeaker[] =>
-  response.map((response) => ({
-    id: response.id,
-    fullName: response.fullName,
-    speakerImage: response.profilePicture,
-    tagLine: response.tagLine,
-    bio: response.bio,
-    sessions: response.sessions,
-    twitterUrl: response.links.filter((link) => link.linkType === "Twitter")[0],
-    linkedInUrl: response.links.filter(
-      (link) => link.linkType === "LinkedIn",
-    )[0],
-  }));
+
